ref #002: create seeder module inside prisma module and remove seeder feature in shared module
