+++ src/prisma/seeder/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './seeder.dto'
