+++ src/prisma/seeder/dto/seeder.dto.ts
@@ -1,0 +1,8 @@
+import { IsEnum } from "class-validator";
+import { TableEnum } from "../../../shared/entities";
+
+
+export class TableSeederDto {
+    @IsEnum(TableEnum)
+    tbl_name: TableEnum;
+}
