+++ src/prisma/seeder/seeder.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { SeederService } from './seeder.service';
+import { SeederController } from './seeder.controller';
+
+@Module({
+  providers: [SeederService],
+  controllers: [SeederController],
+  exports: [SeederService],
+})
+export class SeederModule {}
