--- src/shared/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './seeder.dto'
